backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.1
beanie==1.29.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.0
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.105.0
h11==0.14.0
idna==3.10
ImageHash==4.3.2
lazy-model==0.2.0
motor==3.7.0
numpy==2.2.5
passlib==1.7.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pymongo==4.12.0
pypdfium2==4.30.1
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
PyWavelets==1.8.0
rsa==4.9.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.24.0